requirements.txt
# local package
-e .

# external requirements
websocket-client
pyee